🎁 EDA transcription feature
This commit adds the plugin functionality that allows a user to click a
button on the mirador viewer to see transcriptions of the Emily
Dickinson Archive manuscript images. The plugin allows users to toggle
between viewing the metric line breaks and the physical line breaks, see
and edits made to the manuscript, and view all transcription editions
available for a given manuscript image.